+++ src/loc.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 ///
 /// It must be implemented by application since it is project specific.
 /// See [example implementation](https://github.com/alexfedoseev/steward/tree/master/examples/cli/loc.rs) in the repository.
-pub trait Location: Sized {
+pub trait Location: Sized + Send + Sync {
     /// Returns a location of the root directory of a project.
     ///
     /// It'd be better to call it `root` but this name left vacant because it's used often