+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### UNRELEASED
+### 0.0.4
 - Switch to 2021 Rust edition.
 - Add dependant processes. See [docs](https://docs.rs/steward/latest/steward/dep/index.html).
 - Add [`Cmd::spawn`](https://docs.rs/steward/latest/steward/cmd/struct.Cmd.html#method.spawn) and [`Process::spawn`](https://docs.rs/steward/latest/steward/process/struct.Process.html#method.spawn) methods.